refactored Player and Shredder
